# take cleanPath out of the hot loop