This is probably not significant compared to noise but here it is. 

It's about 2% faster (scripting time), across 10 trials of opening that very large notebook.

Master:
average = 14983ms
stddev   = 1773ms

With fix:
average = 14674ms
stddev   = 1664ms

🤷